cli: compile de manière incrémentale sous VDN

+++ cli/make-in-vdn.sh
@@ -2,18 +2,30 @@
 
 set -eu
 
+VDN_SYSTEM=bigboss
+
 vdn-set-network-dir ~vdn/vdn/networks/demo
-if ! vdn-alive bigboss; then
-  vdn-start -b bigboss
+if ! vdn-alive $VDN_SYSTEM; then
+  vdn-start -b $VDN_SYSTEM
+  until vdn-ssh root@$VDN_SYSTEM : &> /dev/null; do
+    sleep 2
+  done
 fi
-vdn-scp -r src root@bigboss:
-vdn-scp -r Makefile root@bigboss:
-vdn-ssh root@bigboss << EOF
+
+currentDate=$(date +'%F %H:%M:%S')
+vdn-ssh root@$VDN_SYSTEM "date -s '$currentDate' &> /dev/null && date -s '-2 hours'"
+
+rsync -av src -e vdn-ssh root@$VDN_SYSTEM:
+vdn-scp Makefile root@$VDN_SYSTEM:
+vdn-ssh root@$VDN_SYSTEM << EOF
+
 route del default gw 192.168.2.1
 route add default gw 10.0.2.2
 export http_proxy=http://193.49.118.36:8080/
-apt-get install -y nlohmann-json3-dev libcurl4-openssl-dev
+if ! pkg-config nlohmann_json --exists || ! pkg-config libcurl --exists; then
+  apt-get install -y nlohmann-json3-dev libcurl4-openssl-dev
+fi
 make
 EOF
-vdn-scp root@bigboss:oki .
+vdn-scp root@$VDN_SYSTEM:oki .
